#292[WIP] - Add integration tests for redis using dockertest, remove tests using miniredis
